+++ crates/rune/src/compile/unit_builder.rs
@@ -6,8 +6,8 @@
 use crate::ast::Span;
 use crate::collections::HashMap;
 use crate::compile::{
-    Assembly, AssemblyInst, CompileError, CompileErrorKind, IntoComponent, Item, Location,
-    PrivMeta, PrivMetaKind, PrivVariantMeta,
+    Assembly, AssemblyInst, CompileError, CompileErrorKind, Item, Location, PrivMeta, PrivMetaKind,
+    PrivVariantMeta,
 };
 use crate::query::{QueryError, QueryErrorKind};
 use crate::runtime::debug::{DebugArgs, DebugSignature};
@@ -34,8 +34,6 @@ pub enum LinkerError {
 /// Instructions from a single source file.
 #[derive(Debug, Default)]
 pub(crate) struct UnitBuilder {
-    /// Prelude imports.
-    prelude: HashMap<Box<str>, Item>,
     /// The instructions contained in the source file.
     instructions: Vec<Inst>,
     /// Registered re-exports.
@@ -76,47 +74,6 @@ pub(crate) struct UnitBuilder {
 }
 
 impl UnitBuilder {
-    /// Construct a new unit with the default prelude.
-    pub(crate) fn with_default_prelude() -> Self {
-        let mut this = Self::default();
-
-        this.add_prelude("assert_eq", &["test", "assert_eq"]);
-        this.add_prelude("assert", &["test", "assert"]);
-        this.add_prelude("bool", &["bool"]);
-        this.add_prelude("byte", &["byte"]);
-        this.add_prelude("char", &["char"]);
-        this.add_prelude("dbg", &["io", "dbg"]);
-        this.add_prelude("drop", &["mem", "drop"]);
-        this.add_prelude("Err", &["result", "Result", "Err"]);
-        this.add_prelude("file", &["macros", "builtin", "file"]);
-        this.add_prelude("float", &["float"]);
-        this.add_prelude("format", &["fmt", "format"]);
-        this.add_prelude("int", &["int"]);
-        this.add_prelude("is_readable", &["is_readable"]);
-        this.add_prelude("is_writable", &["is_writable"]);
-        this.add_prelude("line", &["macros", "builtin", "line"]);
-        this.add_prelude("None", &["option", "Option", "None"]);
-        this.add_prelude("Object", &["object", "Object"]);
-        this.add_prelude("Ok", &["result", "Result", "Ok"]);
-        this.add_prelude("Option", &["option", "Option"]);
-        this.add_prelude("panic", &["panic"]);
-        this.add_prelude("print", &["io", "print"]);
-        this.add_prelude("println", &["io", "println"]);
-        this.add_prelude("Result", &["result", "Result"]);
-        this.add_prelude("Some", &["option", "Option", "Some"]);
-        this.add_prelude("String", &["string", "String"]);
-        this.add_prelude("stringify", &["stringify"]);
-        this.add_prelude("unit", &["unit"]);
-        this.add_prelude("Vec", &["vec", "Vec"]);
-
-        this
-    }
-
-    /// Clone the prelude.
-    pub(crate) fn prelude(&self) -> &HashMap<Box<str>, Item> {
-        &self.prelude
-    }
-
     /// Convert into a runtime unit, shedding our build metadata in the process.
     ///
     /// Returns `None` if the builder is still in use.
@@ -316,7 +273,6 @@ impl UnitBuilder {
 
     /// Declare a new struct.
     pub(crate) fn insert_meta(&mut self, span: Span, meta: &PrivMeta) -> Result<(), QueryError> {
-        // TODO: Can someone deduplicate this?
         match &meta.kind {
             PrivMetaKind::Unknown { .. } => {
                 let hash = Hash::type_hash(&meta.item.item);
@@ -689,16 +645,6 @@ impl UnitBuilder {
                 );
             }
         }
-    }
-
-    /// Define a prelude item.
-    fn add_prelude<I>(&mut self, local: &str, path: I)
-    where
-        I: IntoIterator,
-        I::Item: IntoComponent,
-    {
-        self.prelude
-            .insert(local.into(), Item::with_crate_item("std", path));
     }
 
     /// Insert and access debug information.
