+++ crates/rune/src/query/mod.rs
@@ -8,12 +8,11 @@ use std::sync::Arc;
 
 use crate::ast;
 use crate::ast::{Span, Spanned};
-use crate::collections::{HashMap, HashSet};
-use crate::compile::ir;
+use crate::collections::{hash_map, HashMap, HashSet};
 use crate::compile::{
-    CaptureMeta, CompileError, CompileErrorKind, CompileVisitor, ComponentRef, ImportStep,
+    ir, CaptureMeta, CompileError, CompileErrorKind, CompileVisitor, ComponentRef, ImportStep,
     IntoComponent, IrBudget, IrCompiler, IrInterpreter, Item, ItemMeta, Location, ModMeta, Names,
-    PrivMeta, PrivMetaKind, PrivStructMeta, PrivTupleMeta, PrivVariantMeta, SourceMeta,
+    Prelude, PrivMeta, PrivMetaKind, PrivStructMeta, PrivTupleMeta, PrivVariantMeta, SourceMeta,
     UnitBuilder, Visibility,
 };
 use crate::hir;
@@ -120,9 +119,14 @@ pub(crate) struct QueryInner {
     modules: HashMap<Item, Arc<ModMeta>>,
 }
 
+/// Query system of the rune compiler.
+///
+/// Once an item is queried for it is queued up for compilation.
 pub(crate) struct Query<'a> {
     /// The current unit being built.
     pub(crate) unit: &'a mut UnitBuilder,
+    /// The prelude in effect.
+    prelude: &'a Prelude,
     /// Cache of constants that have been expanded.
     pub(crate) consts: &'a mut Consts,
     /// Storage associated with the query.
@@ -141,6 +145,7 @@ impl<'a> Query<'a> {
     /// Construct a new compilation context.
     pub(crate) fn new(
         unit: &'a mut UnitBuilder,
+        prelude: &'a Prelude,
         consts: &'a mut Consts,
         storage: &'a mut Storage,
         sources: &'a mut Sources,
@@ -150,6 +155,7 @@ impl<'a> Query<'a> {
     ) -> Self {
         Self {
             unit,
+            prelude,
             consts,
             storage,
             sources,
@@ -163,6 +169,7 @@ impl<'a> Query<'a> {
     pub(crate) fn borrow(&mut self) -> Query<'_> {
         Query {
             unit: self.unit,
+            prelude: self.prelude,
             consts: self.consts,
             storage: self.storage,
             sources: self.sources,
@@ -174,18 +181,21 @@ impl<'a> Query<'a> {
 
     /// Insert the given compile meta.
     pub(crate) fn insert_meta(&mut self, span: Span, meta: PrivMeta) -> Result<(), QueryError> {
-        let item = meta.item.item.clone();
-
         self.visitor.register_meta(meta.info_ref());
 
-        if let Some(existing) = self.inner.meta.insert(item, meta.clone()) {
-            return Err(QueryError::new(
-                span,
-                QueryErrorKind::MetaConflict {
-                    current: meta.info(),
-                    existing: existing.info(),
-                },
-            ));
+        match self.inner.meta.entry(meta.item.item.clone()) {
+            hash_map::Entry::Occupied(e) => {
+                return Err(QueryError::new(
+                    span,
+                    QueryErrorKind::MetaConflict {
+                        current: meta.info(),
+                        existing: e.get().info(),
+                    },
+                ));
+            }
+            hash_map::Entry::Vacant(e) => {
+                e.insert(meta);
+            }
         }
 
         Ok(())
@@ -1240,7 +1250,7 @@ impl<'a> Query<'a> {
             }
         }
 
-        if let Some(item) = self.unit.prelude().get(local) {
+        if let Some(item) = self.prelude.get(local) {
             return Ok(item.clone());
         }
